fix(imbas): resolve FIX-001, FIX-002 from review
